requirements.txt
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.92.0
langchain-google-vertexai>=0.0.1
transformers>=4.30.0
torch>=2.0.0
Pillow>=9.0.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0 